first pass refactoring enemies
